pyproject.toml
[project]
name = "programs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cplex>=22.1.2.0",
    "docplex>=2.30.251",
    "geopandas>=1.1.2",
    "networkx>=3.5",
    "numpy>=2.3.5",
    "ortools>=9.14.6206",
    "pymoo>=0.6.1.6",
]
